Cleans Rev2
