+++ 6Colors/src/edu/isep/sixcolors/model/Config.java
