+++ 6Colors/src/edu/isep/sixcolors/model/GameState.java
