+++ 6Colors/src/edu/isep/sixcolors/model/TileColor.java
@@ -16,49 +16,49 @@ import java.util.Random;
  * </ul>
  */
 public enum TileColor {
-	Blue('B', Color.BLUE),
-	Green('G', Color.GREEN),
-	Red('R', Color.RED),
-	Yellow('Y', Color.YELLOW),
-	Orange('O', Color.ORANGE),
-	Magenta('M', Color.PINK);
-	
-	private static Random random = new Random();
-	
-	/**
-	 * First letter of the color :
-	 */
-	private char initial;
-	private Color color;
-	
-	TileColor(char initial, Color color) {
-		this.setInitial(initial);
-		this.setColor(color);
-	}
-	
-	public static TileColor random() {
-		return TileColor.values()[TileColor.random.nextInt(TileColor.values().length)];
-	}
+    Blue('B', Color.BLUE),
+    Green('G', Color.GREEN),
+    Red('R', Color.RED),
+    Yellow('Y', Color.YELLOW),
+    Orange('O', Color.ORANGE),
+    Magenta('M', Color.PINK);
 
-	
-	/**
-	 * Get the first letter of the color
-	 * @return char : first letter
-	 */
-	public char getInitial() {
-		return initial;
-	}
+    private static Random random = new Random();
 
-	private void setInitial(char initial) {
-		this.initial = initial;
-	}
+    /**
+     * First letter of the color :
+     */
+    private char initial;
+    private Color color;
+
+    TileColor(char initial, Color color) {
+        this.setInitial(initial);
+        this.setColor(color);
+    }
+
+    public static TileColor random() {
+        return TileColor.values()[TileColor.random.nextInt(TileColor.values().length)];
+    }
 
 
-	public Color getColor() {
-		return color;
-	}
+    /**
+     * Get the first letter of the color
+     * @return char : first letter
+     */
+    public char getInitial() {
+        return initial;
+    }
 
-	public void setColor(Color color) {
-		this.color = color;
-	}
+    private void setInitial(char initial) {
+        this.initial = initial;
+    }
+
+
+    public Color getColor() {
+        return color;
+    }
+
+    public void setColor(Color color) {
+        this.color = color;
+    }
 }
