--- 6Colors/src/edu/isep/sixcolors/view/Config.java
@@ -1,7 +1,0 @@
-package edu.isep.sixcolors.view;
-
-/**
- * Created by ohhopi on 26/05/2016.
- */
-public class Config {
-}
