+++ 6Colors/src/edu/isep/sixcolors/view/Window.java
