+++ 6Colors/src/edu/isep/sixcolors/view/game/colorButtons.java
@@ -1,0 +1,20 @@
+package edu.isep.sixcolors.view.game;
+
+
+import edu.isep.sixcolors.model.Game;
+
+import javax.swing.*;
+
+public class colorButtons extends JPanel {
+    private Game game;
+
+    //GetColors
+    //RemoveTakenOnes
+
+    //CreateAButtonForEachFreeOne
+
+    public colorButtons(Game game){
+
+    }
+
+}
