--- 6Colors2/src/edu/isep/sixcolors/view/game/colorButtons.java
@@ -1,7 +1,0 @@
-package edu.isep.sixcolors.view.game;
-
-/**
- * Created by ohhopi on 27/05/2016.
- */
-public class colorButtons {
-}
